+++ app/Http/Controllers/PostsCategoryController.php
@@ -64,7 +64,15 @@ class PostsCategoryController extends Controller
     public function delete($id)
     {
         $posts_category = \App\PostsCategory::find($id);
-        $posts_category->delete($posts_category);
-        return redirect('/admin/kategori')->with('sukses', 'data berhasil di hapus');
+        $post = \App\Posts::where('posts_category_id',$id)->first();
+        if($post)
+        {
+            return redirect('/admin/kategori')->with('gagal', 'data masih ada di tabel posting');
+        }
+        else
+        {
+            $posts_category->delete($posts_category);
+            return redirect('/admin/kategori')->with('sukses', 'data berhasil di hapus');
+        }
     }
 }
